requirements.txt
absl-py==2.3.1
action-msgs==2.0.3
action-tutorials-interfaces==0.33.7
action-tutorials-py==0.33.7
actionlib-msgs==5.3.6
ament-cmake-test==2.5.4
ament-copyright==0.17.3
ament-cppcheck==0.17.3
ament-cpplint==0.17.3
ament-flake8==0.17.3
ament-index-python==1.8.1
ament-lint==0.17.3
ament-lint-cmake==0.17.3
ament-package==0.16.4
ament-pep257==0.17.3
ament-uncrustify==0.17.3
ament-xmllint==0.17.3
angles==1.16.1
asttokens==3.0.1
Box2D==2.3.10
builtin-interfaces==2.0.3
cloudpickle==3.1.2
comm==0.2.3
composition-interfaces==2.0.3
contourpy==1.3.3
cv-bridge==4.1.0
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
demo-nodes-py==0.33.7
diagnostic-msgs==5.3.6
domain-coordinator==0.12.0
example-interfaces==0.12.0
examples-rclpy-executors==0.19.6
examples-rclpy-minimal-action-client==0.19.6
examples-rclpy-minimal-action-server==0.19.6
examples-rclpy-minimal-client==0.19.6
examples-rclpy-minimal-publisher==0.19.6
examples-rclpy-minimal-service==0.19.6
examples-rclpy-minimal-subscriber==0.19.6
executing==2.2.1
Farama-Notifications==0.0.4
filelock==3.20.2
fonttools==4.61.1
fsspec==2025.12.0
geometry-msgs==5.3.6
grpcio==1.76.0
gymnasium==1.2.3
image-geometry==4.1.0
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
interactive-markers==2.5.5
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
laser-geometry==2.7.2
launch==3.4.7
launch-ros==0.26.9
launch-testing==3.4.7
launch-testing-ros==0.26.9
launch-xml==3.4.7
launch-yaml==3.4.7
lifecycle-msgs==2.0.3
logging-demo==0.33.7
map-msgs==2.4.1
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
message-filters==4.11.8
mpmath==1.3.0
nav-msgs==5.3.6
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
osrf-pycommon==2.1.7
packaging==25.0
pandas==2.3.3
parso==0.8.5
pcl-msgs==1.0.0
pendulum-msgs==0.33.7
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
protobuf==6.33.2
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-qt-binding==2.2.2
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qt-dotgraph==2.7.5
qt-gui==2.7.5
qt-gui-cpp==2.7.5
qt-gui-py-common==2.7.5
quality-of-service-demo-py==0.33.7
rcl-interfaces==2.0.3
rclpy==7.1.5
rcutils==6.7.4
resource-retriever==3.4.4
rmw-dds-common==3.1.0
ros2action==0.32.6
ros2bag==0.26.9
ros2bag-mcap-cli==0.26.9
ros2bag-sqlite3-cli==0.26.9
ros2cli==0.32.6
ros2component==0.32.6
ros2doctor==0.32.6
ros2interface==0.32.6
ros2launch==0.26.9
ros2lifecycle==0.32.6
ros2multicast==0.32.6
ros2node==0.32.6
ros2param==0.32.6
ros2pkg==0.32.6
ros2run==0.32.6
ros2service==0.32.6
ros2topic==0.32.6
rosbag2-interfaces==0.26.9
rosbag2-py==0.26.9
rosgraph-msgs==2.0.3
rosidl-adapter==4.6.6
rosidl-cli==4.6.6
rosidl-cmake==4.6.6
rosidl-generator-c==4.6.6
rosidl-generator-cpp==4.6.6
rosidl-generator-py==0.22.2
rosidl-generator-type-description==4.6.6
rosidl-parser==4.6.6
rosidl-pycommon==4.6.6
rosidl-runtime-py==0.13.1
rosidl-typesupport-c==3.2.2
rosidl-typesupport-cpp==3.2.2
rosidl-typesupport-fastrtps-c==3.6.2
rosidl-typesupport-fastrtps-cpp==3.6.2
rosidl-typesupport-introspection-c==4.6.6
rosidl-typesupport-introspection-cpp==4.6.6
rpyutils==0.4.2
rqt-action==2.2.0
rqt-bag==1.5.5
rqt-bag-plugins==1.5.5
rqt-console==2.2.1
rqt-graph==1.5.5
rqt-gui==1.6.1
rqt-gui-py==1.6.1
rqt-msg==1.5.1
rqt-plot==1.4.4
rqt-publisher==1.7.2
rqt-py-common==1.6.1
rqt-py-console==1.2.2
rqt-reconfigure==1.6.2
rqt-service-caller==1.2.1
rqt-shell==1.2.2
rqt-srv==1.2.2
rqt-topic==1.7.3
sensor-msgs==5.3.6
sensor-msgs-py==5.3.6
service-msgs==2.0.3
setuptools==80.9.0
shape-msgs==5.3.6
six==1.17.0
sros2==0.13.4
stack-data==0.6.3
statistics-msgs==2.0.3
std-msgs==5.3.6
std-srvs==5.3.6
stereo-msgs==5.3.6
swig==4.4.1
sympy==1.14.0
teleop-twist-keyboard==2.4.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tf2-geometry-msgs==0.36.14
tf2-kdl==0.36.14
tf2-msgs==0.36.14
tf2-py==0.36.14
tf2-ros-py==0.36.14
tf2-sensor-msgs==0.36.14
tf2-tools==0.36.14
topic-monitor==0.33.7
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
trajectory-msgs==5.3.6
triton==3.5.1
turtlesim==1.8.3
type-description-interfaces==2.0.3
typing_extensions==4.15.0
tzdata==2025.3
unique-identifier-msgs==2.5.0
visualization-msgs==5.3.6
wcwidth==0.2.14
Werkzeug==3.1.4
